split raw project api calls to separate service
