+++ src/test/java/bio/terra/service/resourcemanagement/google/ResourceServiceConnectedTest.java
@@ -43,6 +43,7 @@ public class ResourceServiceConnectedTest {
   @Autowired private GoogleResourceConfiguration resourceConfiguration;
   @Autowired private BufferService bufferService;
   @Autowired private GoogleProjectService projectService;
+  @Autowired private GoogleResourceManagerService resourceManagerService;
   @Autowired private ConnectedOperations connectedOperations;
   @Autowired private ConnectedTestConfiguration testConfig;
   @Autowired private ConfigurationService configService;
@@ -83,13 +84,13 @@ public class ResourceServiceConnectedTest {
             GoogleRegion.DEFAULT_GOOGLE_REGION,
             Map.of("test-name", "resource-service-connected-test"));
 
-    Project project = projectService.getProject(projectId);
+    Project project = resourceManagerService.getProject(projectId);
     assertThat("the project is active", project.getLifecycleState(), equalTo("ACTIVE"));
 
     // TODO check to make sure a steward can complete a job in another test
 
     projectService.deleteGoogleProject(projectResource.getId());
-    project = projectService.getProject(projectId);
+    project = resourceManagerService.getProject(projectId);
     assertThat(
         "the project is not active after delete",
         project.getLifecycleState(),
