+++ src/test/java/bio/terra/service/tabulardata/google/BigQueryPdaoDatasetConnectedTest.java
@@ -30,7 +30,7 @@ import bio.terra.service.dataset.DatasetUtils;
 import bio.terra.service.iam.IamProviderInterface;
 import bio.terra.service.resourcemanagement.BufferService;
 import bio.terra.service.resourcemanagement.ResourceService;
-import bio.terra.service.resourcemanagement.google.GoogleProjectService;
+import bio.terra.service.resourcemanagement.google.GoogleResourceManagerService;
 import bio.terra.service.snapshot.SnapshotDao;
 import com.google.cloud.bigquery.JobInfo;
 import com.google.cloud.bigquery.QueryJobConfiguration;
@@ -73,7 +73,7 @@ public class BigQueryPdaoDatasetConnectedTest {
   @Autowired private SnapshotDao snapshotDao;
   @Autowired private ConnectedOperations connectedOperations;
   @Autowired private ResourceService resourceService;
-  @Autowired private GoogleProjectService projectService;
+  @Autowired private GoogleResourceManagerService resourceManagerService;
   @Autowired private BufferService bufferService;
 
   @MockBean private IamProviderInterface samService;
@@ -335,7 +335,8 @@ public class BigQueryPdaoDatasetConnectedTest {
     dataset.id(UUID.randomUUID());
     ResourceInfo resource = bufferService.handoutResource(dataset.getSecurityClassification());
     String googleProjectId = resource.getCloudResourceUid().getGoogleProjectUid().getProjectId();
-    projectService.addLabelsToProject(googleProjectId, Map.of("test-name", "bigquery-pdao-test"));
+    resourceManagerService.addLabelsToProject(
+        googleProjectId, Map.of("test-name", "bigquery-pdao-test"));
     UUID projectId =
         resourceService.getOrCreateDatasetProject(
             profileModel, googleProjectId, region, dataset.getName(), dataset.getId());
