+++ src/test/java/bio/terra/service/tabulardata/google/BigQueryPdaoTest.java
@@ -35,7 +35,7 @@ import bio.terra.service.iam.IamProviderInterface;
 import bio.terra.service.resourcemanagement.BufferService;
 import bio.terra.service.resourcemanagement.ResourceService;
 import bio.terra.service.resourcemanagement.google.GoogleProjectResource;
-import bio.terra.service.resourcemanagement.google.GoogleProjectService;
+import bio.terra.service.resourcemanagement.google.GoogleResourceManagerService;
 import bio.terra.service.snapshot.Snapshot;
 import bio.terra.service.snapshot.SnapshotDao;
 import bio.terra.service.tabulardata.exception.BadExternalFileException;
@@ -87,7 +87,7 @@ public class BigQueryPdaoTest {
   @Autowired private SnapshotDao snapshotDao;
   @Autowired private ConnectedOperations connectedOperations;
   @Autowired private ResourceService resourceService;
-  @Autowired private GoogleProjectService projectService;
+  @Autowired private GoogleResourceManagerService resourceManagerService;
   @Autowired private BufferService bufferService;
 
   @MockBean private IamProviderInterface samService;
@@ -366,7 +366,8 @@ public class BigQueryPdaoTest {
     ResourceInfo resourceInfo = bufferService.handoutResource(DatasetSecurityClassification.NONE);
 
     String dataProjectId = resourceInfo.getCloudResourceUid().getGoogleProjectUid().getProjectId();
-    projectService.addLabelsToProject(dataProjectId, Map.of("test-name", "bigquery-pdao-test"));
+    resourceManagerService.addLabelsToProject(
+        dataProjectId, Map.of("test-name", "bigquery-pdao-test"));
 
     Snapshot snapshot =
         new Snapshot()
@@ -461,7 +462,8 @@ public class BigQueryPdaoTest {
     dataset.id(UUID.randomUUID());
     ResourceInfo resource = bufferService.handoutResource(dataset.getSecurityClassification());
     String googleProjectId = resource.getCloudResourceUid().getGoogleProjectUid().getProjectId();
-    projectService.addLabelsToProject(googleProjectId, Map.of("test-name", "bigquery-pdao-test"));
+    resourceManagerService.addLabelsToProject(
+        googleProjectId, Map.of("test-name", "bigquery-pdao-test"));
     UUID projectId =
         resourceService.getOrCreateDatasetProject(
             profileModel, googleProjectId, region, dataset.getName(), dataset.getId());
